+++ src/components/details.css
@@ -60,5 +60,5 @@
   font-size: 40px;
   text-align: center;
   color: red;
-  margin-top: 5%;  
+  margin-top: 5%;
 }
